menu pause

+++ Shoot-em_up/Shoot-em_up/ship.h
@@ -5,8 +5,8 @@ struct SDL_Renderer;
 
 class Ship {
 public:
-	int pos_x;
-	int pos_y;
+	float pos_x;
+	float pos_y;
 	SDL_Texture* m_ship;
 	Ship(SDL_Renderer* _renderer);
 	void Render(SDL_Renderer* _renderer);
